backend/requirements.txt
Flask==2.2.3
flask-cors==3.0.10
Pillow==9.5.0
reportlab==3.5
werkzeug==2.2.3
PyPDF2>=3.0.0